requirements.txt
pydantic>=2.0,<3.0
pydantic-settings>=2.0,<3.0
python-json-logger~=3.3
click>=8.0,<9.0
python-dotenv>=1.0,<2.0
requests>=2.31,<3.0
httpx>=0.25,<0.26
pandas>=2.0,<3.0
sqlalchemy>=2.0,<3.0
psycopg2-binary>=2.9,<3.0
alembic>=1.12,<2.0
